Fix: 수정 페이지 이동
TextField, ColorImageCasePage 페이지 이동 기존 페이지 삭제
